apps/zerg/backend/pyproject.toml
[project]
name = "zerg"
version = "0.1.0"
description = "AI Agent Platform Backend"
authors = [
    {name = "David Rose", email = "david.github@drose.io"}
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "python-dotenv>=1.0.0",
    "black",
    "ruff",
    "pytest",
    "fastapi>=0.110.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "openai>=1.65.2",
    "sqlalchemy>=2.0.38",
    "requests>=2.32.3",
    "pytest-xdist>=3.6.1",
    "langsmith>=0.3.14",
    "pytest-asyncio>=0.25.3",
    "apscheduler>=3.11.0",
    "langgraph>=0.3.34",
    "langchain-openai>=0.3.14",
    "python-multipart>=0.0.20",
    "prometheus-client>=0.19.0",
    "cryptography>=44.0.3",
    "httpx[http2]>=0.27.0",
    "jsonschema>=4.23.0",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "simpleeval>=1.0.3",
    "psycopg2-binary>=2.9.0",
    "google-auth>=2.40.3",
    "PyJWT[crypto]>=2.8.0",
    "alembic>=1.13.0",
    "docker>=7.1.0",
    "testcontainers>=4.13.1",
    "psycopg>=3.2.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["zerg"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "E", "F", "Q000"]
ignore = []

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
